controller/requirements.txt
kubernetes>=28.1.0
pyyaml>=6.0
rich>=13.0.0
tabulate>=0.9.0
